# carritodecompras/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Shopping Cart API
  description: API for managing a shopping cart system including products, users, purchases, and detailed purchase history.
  version: v1.0.0
  contact:
    name: Alejandro Granados
    email: alejandro.magb@gmail.com
    url: https://github.com/aledelorbe
...
paths:
...
  /api/users/all:
    get:
      tags:
        - Users
      summary: Get all users
      description: Returns a paginated list of users (showing brief user projections).
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 5
            minimum: 5
      responses:
        '200':
          description: Paginated list of user projections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseUserResponseProjection'
  /api/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Returns a full user record including purchase history details.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The user ID
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    put:
      tags:
        - Users
      summary: Update an existing user
      description: Updates name and lastname of a specific user.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The user ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrors'
        '404':
          description: User not found
        '409':
          description: Username/lastname combination already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'

    delete:
      tags:
        - Users
      summary: Delete a user
      description: Deletes a user by their ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The user ID
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '404':
          description: User not found
  /api/users/name/{name}:
    get:
      tags:
        - Users
      summary: Search users by name
      description: Finds the top 10 users whose name contains the specified string (case-insensitive).
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: The user name snippet
      responses:
        '200':
          description: List of matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseProjection'
  /api/users:
    post:
      tags:
        - Users
      summary: Create a user
      description: Registers a new user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrors'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /api/users/owner/{id}:
    get:
      tags:
        - Users
      summary: Get owner user details
      description: Returns brief owner user details.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The user ID
      responses:
        '200':
          description: User profile found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseProjection'
        '404':
          description: User not found

  # Purchase controller -----------------------------
...
components:
  schemas:
...
    User:
      type: object
      required:
        - name
        - lastname
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "John"
        lastname:
          type: string
          example: "Doe"
        purchases:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseHistory'
    UserResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "John"
        lastname:
          type: string
          example: "Doe"
    UserResponseProjection:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "John"
        lastname:
          type: string
          example: "Doe"
    PurchaseHistory:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        firstImage:
          type: string
          example: "https://example.com/images/headphones.jpg"
        total:
          type: number
          format: double
          example: 399.98
        date:
          type: string
          format: date
          example: "2026-06-20"
...
    PageResponseUserResponseProjection:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserResponseProjection'
        currentPage:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 5
        totalElements:
          type: integer
          format: int64
          example: 45
        totalPages:
          type: integer
          example: 9
        hasNext:
          type: boolean
          example: true
        hasPrevious:
          type: boolean
          example: false
    ValidationErrors:
      type: object
      description: Validation error messages indexed by field name.
      additionalProperties:
        type: string
      example:
        name: "The field name must be obligated"
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
          example: "Error! The user you want to register already exists in the database."
        error:
          type: string
          example: "Conflict"
        status:
          type: integer
          example: 409
        dateTime:
          type: string
          format: date-time
          example: "2026-06-20T12:00:00"